moved GeoMapper to MeshGeoToolsLib
